refactor: move SingleShippingOption and MultipleShippingOptionsForm to its own files
